Store old handler in  file
